Fixes for experimental path
